test(db): isolate workspace migration coverage
